add numberOfWorkersInGroup fn
change type of workgroup.count

+++ engine/worker/Worker.h
@@ -43,6 +43,6 @@ private:
 class WorkerGroup {
 public:
     void * id = nullptr;
-    int count = 0;
+    uint16_t count = 0;
     Worker::Fn onComplete = nullptr;
 };
